ai/requirements.txt
flask
crewai
google-generativeai 
langchain
dotenv
